Move slash-commands into own file
